fixed arithmetic

+++ arithmetic.cpp
@@ -364,7 +364,7 @@ int main () {
 	}
 	else {
 		std::cout << "Answer:" << std::endl;
-		std::cout << "\tDouble:      " << ans << " in " << global.reset() << " seconds." << std::endl;
+		std::cout << "\tDouble:      " << (double)ans << " in " << global.reset() << " seconds." << std::endl;
 		std::cout << "\tHexadecimal: " << ans.toHex() << " in " << global.reset() << " seconds." << std::endl;
 		std::cout << "\tDecimal:     " << ans.toDec() << " in " << global.reset() << " seconds." << std::endl;
 		std::cout << "\tOctal:       " << ans.toOct() << " in " << global.reset() << " seconds." << std::endl;
